SAS Assignment and Labs
This commit adds a README.md file containing information about the Enterprise Analytics Labs. It includes details about the labs conducted as part of the course, such as the objectives, tasks performed, and analysis conducted. The README provides a comprehensive overview of the labs, helping users understand the scope and purpose of each exercise. And SAS Assignment course on applied-analytics-using-sas-enterprise-miner.